a-softmax

+++ A_softmax.py
@@ -30,7 +30,7 @@ def func_thelta(cos_thelta, m, batch_size):
         for j in range(m):
             idx = m - j - 1
             k[i] = tf.cond(tf.greater_equal(cos_thelta[i], L_constant[idx]), lambda: idx, lambda: 0)
-    K = tf.Variable(k)
+    K =tf.cast(tf.Variable(k), tf.int32)
     if m == 2:
         cos_m_thelta = 2 * cos_thelta ** 2 - 1
     elif m == 3:
